Rework auth-related UI hoisting to shared where sensible
