+++ packages/user/FractalCore/ui/src/lib/queryKeys.ts
@@ -18,9 +18,7 @@ const QueryKey = {
     guild: (guild: OptionalAccount) => ["guild", guild] as const,
     currentUser: () => ["currentUser"] as const,
     chainId: () => ["chainId"] as const,
-    connectedAccounts: () => ["connectedAccounts"] as const,
     createGuild: () => ["createGuild"] as const,
-    selectAccount: () => ["selectAccount"] as const,
     userAccount: (account: string | undefined) => ["userAccount", account],
     fractal: (account: OptionalAccount) => ["fractal", account] as const,
     branding: () => ["branding"] as const,
